fix(FourierLayers): Fix model - fix x_ft slicing bug - fix irfft normalization - Other minor fixes refactor(training): update training parameters.
